Force the bounding boxes predictions of img_inference to be int type

+++ ai_utils/detectors/Yolov8Inference.py
@@ -57,7 +57,7 @@ class Yolov8Inference(DetectorInterface):
                             inference_out[cls]['boxes'] = []
                             inference_out[cls]['masks'] = []
                         inference_out[cls]['scores'].append(scores[idx])
-                        inference_out[cls]['boxes'].append(boxes[idx])
+                        inference_out[cls]['boxes'].append(boxes[idx].astype(int))
                         inference_out[cls]['masks'].append(masks[idx])
 
         return inference_out
